pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "redis>=6.4.0",
    "requests>=2.32.4",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
